Moved utility function to ast-util
